Abstract webpack config and remove comma
