+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "stage": 0,
+  "stage": 0
 }
