+++ tasks/cfg/webpack-config.js
@@ -1,0 +1,36 @@
+var path = require('path');
+var ExtractTextPlugin = require('extract-text-webpack-plugin');
+
+var ROOT_PATH = path.resolve(__dirname);
+
+module.exports = {
+  entry: {
+    app: path.resolve(ROOT_PATH, 'src/app/client/index.js'),
+    styles: path.resolve(ROOT_PATH, 'src/styles/app.styl')
+  },
+  output: {
+    path: path.resolve(ROOT_PATH, 'build/public'),
+    filename: 'app.js'
+  },
+  module: {
+    loaders: [
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        loaders: ['babel-loader', 'eslint-loader']
+      },
+      {
+        test: /\.styl$/,
+        exclude: /node_modules/,
+        loader: ExtractTextPlugin.extract('style', 'css!stylus')
+      },
+      {
+        test: /\.json$/,
+        loader: 'json-loader'
+      }
+    ]
+  },
+  plugins: [
+    new ExtractTextPlugin('styles.css')
+  ]
+};
